chore(code): move `hardwaremapping` package from `node` to `cluster` in `fwprovider`
