+++ fwprovider/cluster/hardwaremapping/datasource_usb.go
