+++ fwprovider/cluster/hardwaremapping/errors.go
