+++ fwprovider/cluster/hardwaremapping/models.go
