+++ fwprovider/cluster/hardwaremapping/resource_dir.go
