+++ fwprovider/cluster/hardwaremapping/resource_hardware_mapping_test.go
