+++ fwprovider/cluster/hardwaremapping/resource_usb.go
