+++ fwprovider/cluster/hardwaremapping/shared.go
